extract QuickFindToolBar from QuickFindDialog disable automatic QuickFindDialog creation when SyntaxpainConfiguration has null dialog factory
