+++ build.gradle
@@ -10,8 +10,10 @@ repositories {
     mavenCentral()
 }
 
+final ENV = System.getenv()
+
 group = 'org.quiltmc'
-version = '0.1.5'
+version = "0.1.5${ENV.GITHUB_ACTIONS ? '' : '+local'}"
 
 // exclude auto-generated jflex lexer from checkstyle
 tasks.withType(Checkstyle).configureEach {
@@ -31,7 +33,6 @@ publishing {
     repositories {
         mavenLocal()
 
-        def ENV = System.getenv()
         if (ENV.MAVEN_URL) {
             maven {
                 url ENV.MAVEN_URL
